Implementacion de CRUD de imagenes del producto
